Fix CAN filter range and set_counts type cast
Expanded CAN filter range in MX_FDCAN1_Init to catch all API IDs. Corrected set_counts to cast counts to int32_t for singleTurnSensorCounts. Cleaned up SET_POSITION_API_ID handling in HAL_FDCAN_RxFifo0Callback for proper position setting. Updated build artifacts accordingly.

+++ Encoder4 june 2025/Encoder4/Core/Src/MT6835.c
@@ -232,7 +232,7 @@ void set_counts(int64_t counts)
 {
     // Set the multi-turn counts and reset single-turn counts
     multiTurnSensorCounts = counts; // Extract multi-turn part
-    singleTurnSensorCounts = counts; // Extract single-turn part
+    singleTurnSensorCounts = (int32_t)counts; // Extract single-turn part
 }
 
 void set_encoder_direction(int dir) {
